refactor: separate subtree discovery version implementations
